Group: consolidate common Mapping-like code in Group
